refactor: use path as entry ID
This is more efficient, for when the entry ID was the denote ID, determining the
entry ID to the denote file path was an incredibly expensive computation. I am
not sure what pitfalls this may lead to. For instance, I know that file paths
present in a buffer can change while that buffer is opened, therefore
necessitating a refresh of the buffer. In any case, this is a much, much quicker
computation for certain actions.